# rest/docs/swagger.yaml
basePath: /
definitions:
...
  api.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  api.PagedResults:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total_records:
        type: integer
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  model.Shuttle:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      handle:
        type: string
      host:
        type: string
      id:
        type: integer
      last_connection:
        type: string
      open:
        type: boolean
      peer_id:
        type: string
      priority:
        type: integer
      private:
        type: boolean
      token:
        type: string
      updated_at:
        type: string
    type: object
  model.StorageMiner:
    properties:
      address:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      location:
        type: string
      name:
        type: string
      owner:
        type: integer
      suspended:
        type: boolean
      suspended_reason:
        type: string
      updated_at:
        type: string
      version:
        type: string
    type: object
...
host: localhost:3030
info:
  contact:
    email: me@me.com
    name: Me
    url: http://me.com/terms.html
  description: Sample CRUD objects-api for estuary db
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Sample CRUD objects-api for estuary db
  version: "1.0"
paths:
...
  /shuttles:
    get:
      consumes:
      - application/json
      description: GetAllShuttles is a handler to get a slice of record(s) from shuttles
        table in the estuary database
      parameters:
      - description: page requested (defaults to 0)
        in: query
        name: page
        type: integer
      - description: number of records in a page  (defaults to 20)
        in: query
        name: pagesize
        type: integer
      - description: db sort order column
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.PagedResults'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Shuttle'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Get list of Shuttles
      tags:
      - Shuttles
    post:
      consumes:
      - application/json
      description: add to add a single record to shuttles table in the estuary database
      parameters:
      - description: Add Shuttles
        in: body
        name: Shuttles
        required: true
        schema:
          $ref: '#/definitions/model.Shuttle'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Shuttle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Add an record to shuttles table
      tags:
      - Shuttles
  /shuttles/{argID}:
    delete:
      consumes:
      - application/json
      description: Delete a single record from shuttles table in the estuary database
      parameters:
      - description: id
        in: path
        name: argID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/model.Shuttle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Delete a record from shuttles
      tags:
      - Shuttles
    get:
      consumes:
      - application/json
      description: GetShuttles is a function to get a single record from the shuttles
        table in the estuary database
      parameters:
      - description: id
        in: path
        name: argID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Shuttle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: ErrNotFound, db record for id not found - returns NotFound
            HTTP 404 not found error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Get record from table Shuttles by  argID
      tags:
      - Shuttles
    put:
      consumes:
      - application/json
      description: Update a single record from shuttles table in the estuary database
      parameters:
      - description: id
        in: path
        name: argID
        required: true
        type: integer
      - description: Update Shuttles record
        in: body
        name: Shuttles
        required: true
        schema:
          $ref: '#/definitions/model.Shuttle'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Shuttle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Update an record in table shuttles
      tags:
      - Shuttles
  /storageminers:
    get:
      consumes:
      - application/json
      description: GetAllStorageMiners is a handler to get a slice of record(s) from
        storage_miners table in the estuary database
      parameters:
      - description: page requested (defaults to 0)
        in: query
        name: page
        type: integer
      - description: number of records in a page  (defaults to 20)
        in: query
        name: pagesize
        type: integer
      - description: db sort order column
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.PagedResults'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.StorageMiner'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Get list of StorageMiners
      tags:
      - StorageMiners
    post:
      consumes:
      - application/json
      description: add to add a single record to storage_miners table in the estuary
        database
      parameters:
      - description: Add StorageMiners
        in: body
        name: StorageMiners
        required: true
        schema:
          $ref: '#/definitions/model.StorageMiner'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.StorageMiner'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Add an record to storage_miners table
      tags:
      - StorageMiners
  /storageminers/{argID}:
    delete:
      consumes:
      - application/json
      description: Delete a single record from storage_miners table in the estuary
        database
      parameters:
      - description: id
        in: path
        name: argID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/model.StorageMiner'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Delete a record from storage_miners
      tags:
      - StorageMiners
    get:
      consumes:
      - application/json
      description: GetStorageMiners is a function to get a single record from the
        storage_miners table in the estuary database
      parameters:
      - description: id
        in: path
        name: argID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.StorageMiner'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: ErrNotFound, db record for id not found - returns NotFound
            HTTP 404 not found error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Get record from table StorageMiners by  argID
      tags:
      - StorageMiners
    put:
      consumes:
      - application/json
      description: Update a single record from storage_miners table in the estuary
        database
      parameters:
      - description: id
        in: path
        name: argID
        required: true
        type: integer
      - description: Update StorageMiners record
        in: body
        name: StorageMiners
        required: true
        schema:
          $ref: '#/definitions/model.StorageMiner'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.StorageMiner'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Update an record in table storage_miners
      tags:
      - StorageMiners
...
swagger: "2.0"